restructure CLI
