+++ pacco/cli/commands/binary.py
